+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@egjs/react-flicking": "^4.11.0",
     "@ionic/cli": "^7.1.1",
     "@splidejs/splide": "^4.1.4",
     "@web3modal/ethereum": "^2.7.1",
@@ -19,8 +20,11 @@
     "next": "13.4.19",
     "postcss": "8.4.28",
     "react": "18.2.0",
+    "react-confetti": "^6.1.0",
     "react-dom": "18.2.0",
     "react-icons": "^4.10.1",
+    "react-slick": "^0.29.0",
+    "react-use": "^17.4.0",
     "tailwindcss": "3.3.3",
     "typewriter-effect": "^2.20.1",
     "viem": "^1.9.2",
